% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{connection_config}
\alias{connection_config}
\title{Read configuration values for a connection}
\usage{
connection_config(sc, prefix = NULL)
}
\arguments{
\item{sc}{\code{spark_connection}}

\item{prefix}{Optional prefix to read parameters for
(e.g. \code{spark.context.}, \code{spark.sql.}, etc.)}
}
\value{
Named list of config parameters (note that if a prefix was
 specified then the names will not include the prefix)
}
\description{
Read configuration values for a connection
}

