% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoke.R
\name{invoke}
\alias{invoke}
\alias{invoke_new}
\alias{invoke_static}
\title{Execute a method on a remote Java object}
\usage{
invoke(jobj, method, ...)

invoke_static(connection, class, method, ...)

invoke_new(connection, class, ...)
}
\arguments{
\item{jobj}{Java object to execute method on.}

\item{method}{Name of method to execute.}

\item{...}{Unused (future expansion)}

\item{connection}{Connection to execute on.}

\item{class}{Class to execute static method on.}
}
\description{
Execute a method on a remote Java object
}

