% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magrittr.R
\name{\%>\%}
\alias{\%>\%}
\title{magrittr forward-pipe operator}
\arguments{
\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using the magrittr semantics.}
}
\description{
Pipe an object forward into a function or call expression.
}
\keyword{internal}

