% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{read_config}
\alias{read_config}
\title{Read configuration values}
\usage{
read_config(config, master, prefix = NULL)
}
\arguments{
\item{config}{List with configuration values}

\item{master}{Master node}

\item{prefix}{Optional prefix to read parameters for
(e.g. \code{spark.context.}, \code{spark.sql.}, etc.)}
}
\value{
Named list of config parameters (note that if a prefix was
 specified then the names will not include the prefix)
}
\description{
Read configuration values
}

