% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{read_shell_file}
\alias{read_shell_file}
\title{Read the shell file from the Spark R Backend}
\usage{
read_shell_file(shell_file)
}
\arguments{
\item{shell_file}{Shell file to read}
}
\value{
List with \code{backendPort}, \code{monitorPort}, and
  \code{rLibraryPath}
}
\description{
Read the shell file and extract the backend port, monitor port,
and R library path.
}
\keyword{internal}

