% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{spark_connection_create}
\alias{spark_connection_create}
\title{Create a spark_connection object.}
\usage{
spark_connection_create(spark_context, hive_context, backend, monitor)
}
\arguments{
\item{spark_context}{Instance of SparkContext object}

\item{hive_context}{Instance of HiveContext object}

\item{backend}{R socket connection to backend}

\item{monitor}{R socket connection for monitor}
}
\value{
Object of class \code{spark_connection}.
}
\description{
Create a \code{spark_connection} based on the specified backend
and monitor sockets.
}
\details{
This connection can be passed to the
\code{\link{invoke_new}} and \code{\link{invoke_static}}
functions.
}
\seealso{
\code{\link{spark_connection}}
}
\keyword{internal}

