% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{spark_version.character}
\alias{spark_version.character}
\title{Version of Spark for a connection}
\usage{
\method{spark_version}{character}(x)
}
\arguments{
\item{x}{Path to SPARK_HOME}
}
\description{
Version of Spark for a connection
}

