% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{spark_version.spark_connection}
\alias{spark_version.spark_connection}
\title{Version of Spark for a connection}
\usage{
\method{spark_version}{spark_connection}(x)
}
\arguments{
\item{x}{\code{spark_connection}}
}
\description{
Version of Spark for a connection
}

