% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{sparkapi_connection}
\alias{sparkapi_connection}
\title{Get the sparkapi_connection associated with an object}
\usage{
sparkapi_connection(x, ...)
}
\arguments{
\item{x}{Object to extract connection from}

\item{...}{Reserved for future use}
}
\value{
A \code{sparkapi_connection} object that can be passed to
  \code{\link{sparkapi_invoke_new}} and \code{\link{sparkapi_invoke_static}}.
}
\description{
S3 method to get the sparkapi_connection associated with objects of
various types.
}
\seealso{
\code{\link{sparkapi_connection_create}}
}

