% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{sparkapi_connection_create}
\alias{sparkapi_connection_create}
\title{Create a sparkapi_connection object.}
\usage{
sparkapi_connection_create(backend, monitor)
}
\arguments{
\item{backend}{R socket connection to backend}

\item{monitor}{R socket connection for monitor}
}
\value{
Object of class \code{sparkapi_connection}.
}
\description{
Create a \code{sparkapi_connection} based on the specified backend
and monitor sockets.
}
\details{
This connection can be passed to the
\code{\link{sparkapi_invoke_new}} and \code{\link{sparkapi_invoke_static}}
functions.
}
\seealso{
\code{\link{sparkapi_connection}}
}

