% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{sparkapi_dataframe}
\alias{sparkapi_dataframe}
\title{Get the Spark DataFrame associated with an object}
\usage{
sparkapi_dataframe(x, ...)
}
\arguments{
\item{x}{Object to get DataFrame from}

\item{...}{Reserved for future use}
}
\value{
Reference to DataFrame
}
\description{
S3 method to get the Spark DataFrame associated with objects of
various types.
}

