% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency.R
\name{sparkapi_dependencies_from_extension}
\alias{sparkapi_dependencies}
\alias{sparkapi_dependencies_from_extension}
\alias{sparkapi_dependencies_from_extensions}
\title{Get the sparkapi dependencies for extension packages}
\usage{
sparkapi_dependencies_from_extension(config, extension)

sparkapi_dependencies_from_extensions(config, extensions)
}
\arguments{
\item{config}{R list containing configuration data}

\item{extension}{Name of extension package}

\item{extensions}{Names of extension packages.}
}
\value{
List of objects of type \code{spark_dependency}.
}
\description{
Call the \code{sparkapi_dependencies} function of the specified
package to extract it's dependencies.
}
\keyword{internal}

