% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency.R
\name{sparkapi_dependency}
\alias{sparkapi_dependency}
\title{Define a Spark API dependency}
\usage{
sparkapi_dependency(jars, packages)
}
\arguments{
\item{jars}{Character vector of full paths to JAR files}

\item{packages}{Character vector of Spark packages names}
}
\value{
An object of type `sparkapi_dependency`
}
\description{
Define a Spark API dependency consisting of a set of custom JARs and Spark packages.
}

