% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{sparkapi_hive_context}
\alias{sparkapi_hive_context}
\title{Get the HiveContext associated with a connection}
\usage{
sparkapi_hive_context(connection)
}
\arguments{
\item{connection}{Connection to get HiveContext from}
}
\value{
Reference to HiveContext
}
\description{
Get the HiveContext \code{spark_jobj} associated with a
\code{sparkapi_connection}
}

