% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobj.R
\name{sparkapi_jobj}
\alias{sparkapi_jobj}
\title{Get the sparkapi_jobj associated with an object}
\usage{
sparkapi_jobj(x, ...)
}
\arguments{
\item{x}{Object to extract jobj from}

\item{...}{Reserved for future use}
}
\value{
A \code{sparkapi_jobj} object that can be passed to
  \code{\link{sparkapi_invoke}}.
}
\description{
S3 method to get the sparkapi_jobj associated with objects of
various types.
}
\seealso{
\code{\link{sparkapi_invoke}}
}

