% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{sparkapi_spark_context}
\alias{sparkapi_spark_context}
\title{Get the SparkContext associated with a connection}
\usage{
sparkapi_spark_context(connection)
}
\arguments{
\item{connection}{Connection to get SparkContext from}
}
\value{
Reference to SparkContext
}
\description{
Get the SparkContext \code{spark_jobj} associated with a
\code{sparkapi_connection}
}

