% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoke.R
\name{sparkapi_stop_backend}
\alias{sparkapi_stop_backend}
\title{Stop the RBackend that services RPC request}
\usage{
sparkapi_stop_backend(connection)
}
\arguments{
\item{connection}{Connection to stop the backend for}
}
\description{
Stop the RBackend that services RPC request
}
\keyword{internal}

