% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell.R
\name{start_shell}
\alias{start_shell}
\title{Start the Spark R Shell}
\usage{
start_shell(spark_home, env = NULL, args = NULL, extensions = NULL)
}
\arguments{
\item{env}{character vector of name=value strings to set environment
    variables.}

\item{args}{a character vector of arguments to \command{command}.}

\item{Extension}{packages to include dependencies for
(see \code{\link{spark_dependency}}).}
}
\description{
Start the Spark R Shell
}

