% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoke.R
\name{unknown_error_handler}
\alias{unknown_error_handler}
\title{Get or set the current unknown error handler}
\usage{
unknown_error_handler(handler = NULL)
}
\arguments{
\item{handler}{A function which accepts a \code{spark_connection}
and returns a single-element character vector. Pass \code{NULL}
to read the current value (if any).}
}
\value{
The current handler if \code{NULL} is passed and the
  previously installed handler when a new handler is passed.
}
\description{
Get or set the current unknown error handler. This function
is called when the backend fails to report an error message.
}
\keyword{internal}

