% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{spark_version}
\alias{spark_version}
\title{Version of Spark for a connection}
\usage{
spark_version(sc)
}
\arguments{
\item{sc}{\code{spark_connection}}
}
\description{
Version of Spark for a connection
}

