% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{spark_version}
\alias{spark_version}
\title{Version of Spark for a connection or SPARK_HOME}
\usage{
spark_version(x)
}
\arguments{
\item{x}{\code{spark_connection} or path to SPARK_HOME}
}
\description{
Version of Spark for a connection or SPARK_HOME
}

