% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell.R
\name{start_shell}
\alias{start_shell}
\title{Start the Spark R Shell}
\usage{
start_shell(spark_home, jars = NULL, packages = NULL, extensions = NULL,
  environment = NULL, shell_args = NULL)
}
\arguments{
\item{Extension}{packages to include dependencies for
(see \code{\link{spark_dependency}}).}
}
\description{
Start the Spark R Shell
}

